Crate ort_sys

Modules

stub

Structs

OrtAllocator

\brief Memory allocation interface

OrtApi

\brief The C API

OrtApiBase

\brief The helper interface to get the right version of OrtApi

OrtArenaCfg

OrtCANNProviderOptions

OrtCUDAProviderOptions

\brief CUDA Provider Options

OrtCUDAProviderOptionsV2

OrtCheckpointState

OrtCompileApi

OrtCustomHandleType

OrtCustomOp

OrtCustomOpDomain

OrtDeviceEpIncompatibilityDetails

OrtDmlApi

OrtDmlDeviceOptions

OrtDnnlProviderOptions

OrtEnv

OrtEnvCreationOptions

OrtEpApi

OrtEpAssignedNode

OrtEpAssignedSubgraph

OrtEpDevice

OrtExternalInitializerInfo

OrtExternalMemoryHandle

OrtExternalResourceImporter

OrtExternalSemaphoreHandle

OrtGraph

OrtHardwareDevice

OrtInteropApi

OrtIoBinding

OrtKernelContext

OrtKernelInfo

OrtKeyValuePairs

OrtLogger

OrtLoraAdapter

OrtMIGraphXProviderOptions

\brief MIGraphX Provider Options

OrtMapTypeInfo

OrtMemoryInfo

OrtModel

OrtModelCompilationOptions

OrtModelEditorApi

OrtModelMetadata

OrtNode

OrtNvTensorRtRtxProviderOptions

OrtOp

OrtOpAttr

OrtOpenVINOProviderOptions

\brief OpenVINO Provider Options

OrtOptionalTypeInfo

OrtPrepackedWeightsContainer

OrtROCMProviderOptions

\brief ROCM Provider Options

OrtRunOptions

OrtSequenceTypeInfo

OrtSession

OrtSessionOptions

OrtShapeInferContext

OrtStatus

OrtStatusPtr

OrtSyncStream

OrtTensorRTProviderOptions

\brief TensorRT Provider Options

OrtTensorRTProviderOptionsV2

OrtTensorTypeAndShapeInfo

OrtThreadPoolCallbacksConfig

OrtThreadPoolParams

OrtThreadingOptions

OrtTrainingApi

OrtTrainingSession

OrtTypeInfo

OrtValue

OrtValueInfo

Enums

ExecutionMode

GraphOptimizationLevel

\brief Graph optimization level

ONNXTensorElementDataType

Copied from TensorProto::DataType Currently, Ort doesn’t support complex64, complex128

ONNXType

OrtAllocatorType

OrtCompiledModelCompatibility

OrtCudnnConvAlgoSearch

\brief Algorithm to use for cuDNN Convolution Op

OrtCustomOpInputOutputCharacteristic

OrtDeviceMemoryType

OrtDmlDeviceFilter

OrtDmlPerformancePreference

OrtErrorCode

OrtExecutionProviderDevicePolicy

OrtHardwareDeviceType

OrtLanguageProjection

\brief Language projection identifiers /see OrtApi::SetLanguageProjection

OrtLoggingLevel

\brief Logging severity levels

OrtMemType

\brief Memory types for allocated memory, execution provider specific types should be extended in each provider.

OrtMemoryInfoDeviceType

\brief This mimics OrtDevice type constants so they can be returned in the API

OrtOpAttrType

OrtPropertyType

OrtSparseFormat

OrtSparseIndicesFormat

c_void

Equivalent to C’s void type when used as a pointer.

Constants

ORT_API_VERSION

Functions

OrtGetApiBase⚠

\brief The Onnxruntime library’s entry point to access the C API

OrtSessionOptionsAppendExecutionProvider_CUDA⚠

OrtSessionOptionsAppendExecutionProvider_Dnnl⚠

OrtSessionOptionsAppendExecutionProvider_MIGraphX⚠

OrtSessionOptionsAppendExecutionProvider_ROCM⚠

OrtSessionOptionsAppendExecutionProvider_Tensorrt⚠

Type Aliases

EpSelectionDelegate

OrtCustomCreateThreadFn

\brief Ort custom thread creation function

OrtCustomJoinThreadFn

\brief Custom thread join function

OrtCustomThreadHandle

OrtGetInitializerLocationFunc

OrtLoggingFunction

OrtThreadPoolWorkAbandonFn

OrtThreadPoolWorkEnqueueFn

OrtThreadPoolWorkStartFn

OrtThreadPoolWorkStopFn

OrtThreadWorkerFn

\brief Thread work loop function

OrtWriteBufferFunc

RunAsyncCallbackFn

\brief Callback function for RunAsync

c_char

Equivalent to C’s char type.

c_int

Equivalent to C’s signed int (int) type.

c_ulong

Equivalent to C’s unsigned long type.

c_ulonglong

Equivalent to C’s unsigned long long type.

c_ushort

Equivalent to C’s unsigned short type.

ortcharDeprecated

os_char